back_agridecis/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
Pillow>=10.0,<12.0
requests>=2.31,<3.0
ultralytics>=8.2,<9.0
